CU-2ynbcdm - WEB-005: Working Login - JWT -Little clean up
