+++ src/main/java/com/uz/shop/animal/world/models/UserType.java
@@ -1,4 +1,4 @@
-package com.uz.shop.animal.world.security.user;
+package com.uz.shop.animal.world.models;
 
 public enum UserType {
     ADMIN,
