+++ src/main/java/com/uz/shop/animal/world/security/config/SecurityConfig.java
@@ -1,25 +1,20 @@
 package com.uz.shop.animal.world.security.config;
 
-import com.uz.shop.animal.world.security.user.UserService;
+import com.uz.shop.animal.world.security.config.filter.CustomAuthenticationFilter;
+import com.uz.shop.animal.world.security.config.filter.CustomAuthorizationFilter;
+import com.uz.shop.animal.world.services.UserService;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.OAuth2AuthorizationServerConfiguration;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.oauth2.server.authorization.client.RegisteredClientRepository;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-import static org.springframework.security.config.Customizer.withDefaults;
 
 @Configuration
 @AllArgsConstructor
@@ -42,7 +37,7 @@ public class SecurityConfig {
 
         http.csrf().disable();
         http.authorizeRequests().antMatchers("/api/v*/register/**").permitAll();
-        http.authorizeRequests().antMatchers("api/v*/token/refresh/**").permitAll();
+        http.authorizeRequests().antMatchers("/api/v*/token/refresh/**").permitAll();
         http.authorizeRequests().antMatchers("/api/v*/recovery/**").permitAll();
         http.authorizeRequests().antMatchers("/api/v*/test");
         http.authorizeRequests().anyRequest().authenticated();
