+++ src/main/java/com/uz/shop/animal/world/security/config/filter/CustomAuthorizationFilter.java
@@ -1,11 +1,11 @@
-package com.uz.shop.animal.world.security.config;
+package com.uz.shop.animal.world.security.config.filter;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.JWTVerifier;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.uz.shop.animal.world.security.AuthorizationToken;
+import com.uz.shop.animal.world.services.AuthorizationService;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
@@ -35,7 +35,7 @@ public class CustomAuthorizationFilter extends OncePerRequestFilter {
             if(authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
                 try {
                     String token = authorizationHeader.substring("Bearer ".length());
-                    Algorithm algorithm = AuthorizationToken.getAlgorithm();
+                    Algorithm algorithm = AuthorizationService.getAlgorithm();
                     JWTVerifier verifier = JWT.require(algorithm).build();
                     DecodedJWT decodedJWT = verifier.verify(token);
                     String username = decodedJWT.getSubject();
