+++ src/main/java/com/uz/shop/animal/world/services/email/EmailSender.java
@@ -1,4 +1,4 @@
-package com.uz.shop.animal.world.email;
+package com.uz.shop.animal.world.services.email;
 
 public interface EmailSender {
     void send(String email, String body);
